+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Added
 - Add QObject/QML wrapper for stores to access a store directly in QML.
 - Generate store variable changed callbacks.
 - Allow string variables to be initialized.
+- Generate store meta data in ``doc/<store>Meta.py``
 
 .. _Unreleased: https://github.com/DEMCON/libstored/compare/v1.1.0...HEAD